Remade collection manipulation functions for more efficiency.
The heights are now kept ordered and the laser consult takes only O(log number of registered buildings) time. On the other hand, the insertion is no longer O(1), since it has to keep the collections ordered property. At the moment, I couldn't determine if the implemented solution takes O(log number of registered buildings) or O(number of registered buildings), since it builds up onto the native functions "split-at", "concat" and "vec", which I still couldn't find the time complexity.

Apart from that, even if the insertion is linear, the current solution might be better than the previous if the laser consult interaction is expected to happen more frequently than the building registering.